requirements.txt
duckdb>=1.0.0
shapely>=2.0.0
pyyaml>=6.0
click>=8.0
h3>=4.0.0
rich>=13.0.0
streamlit
leafmap
folium
